refact: refact DTO name
